IDAM-307 Refactor claims and email API models
