--- NICE.Identity.Authorisation.WebAPI/APIModels/Responses/ErrorDetail.cs
@@ -1,9 +1,0 @@
-﻿namespace NICE.Identity.Authorisation.WebAPI.ApiModels.Responses
-{
-    public class ErrorDetail
-    {
-        public string ErrorCode { get; set; }
-
-        public string ErrorMessage { get; set; }
-    }
-}
